Add ForEach()

+++ NetFabric.Hyperlinq/Aggregation/Count/Count.ReadOnlySpan.cs
@@ -25,7 +25,7 @@ namespace NetFabric.Hyperlinq
         }
 
         [Pure]
-        public static int Count<TSource>(this ReadOnlySpan<TSource> source, Func<TSource, long, bool> predicate)
+        public static int Count<TSource>(this ReadOnlySpan<TSource> source, Func<TSource, int, bool> predicate)
         {
             if (predicate is null) ThrowHelper.ThrowArgumentNullException(nameof(predicate));
